Fix the order in which we apply the datagrid queryset data augmentation.
The new support for queryset augmentation for data display purposes
didn't work quite right. It applied to the initial query, but we'd then
end up generating a new query based on the IDs returned from that query.
That new query was augmented the old-fashioned way, but not the new way.

This change fixes that up so that we properly augment the final
data-fetching queryset.

Testing Done:
Unit tests pass.

Tested this with some in-progress datagrid fixes in Review Board.

Reviewed at https://reviews.reviewboard.org/r/13421/