Add support for matching join types in equeries.
This introduces a `join_types` field for equeries, which helps
instrument join types (such as INNER vs. OUTER JOINs). That can be
useful in ensuring that queries are joining in a way that the developers
expect, which can have an impact on results.

This is off by default, for compatibility. If a mismatch would have been
reported, then a warning will be issued advising the developer to opt in
and that it will be made the default in a future version.

There's also a fix for formatting `set`s, ensuring they're both sorted
and pretty-printed.

Testing Done:
Unit tests pass.

Made use of this in some in-progress tests.

Reviewed at https://reviews.reviewboard.org/r/13424/