Move JSONField test model to a higher scope.
We had a bunch of unit tests for the JSONField implementation that were
defining a test model inside the test case itself. This led to warnings
about the model being reregistered. Django doesn't provide a way to
unregister a model at runtime, so our best option for fixing this is to
make sure that the model is only registered once, by moving it up to the
file scope.

Testing Done:
Ran unit tests and saw that the warning was gone.

Reviewed at https://reviews.reviewboard.org/r/13489/