Switch a lot of conditional behavior over from settings.DEBUG to PRODUCTION.
We have a lot of stuff that needs to behave differently in a development
environment than in production. We initially conditioned all of these on
the `DEBUG` setting, but a lot of users tend to turn that on when they
encounter problems, and then sometimes forget that they did so.

On the Review Board side, we've moved these things behind a new settings
key, `PRODUCTION`. This allows people to set `DEBUG = True` and get more
thorough logging, while not completely breaking their system. This
change brings that same key to Djblets.

Testing Done:
- Ran unit tests.
- Smoke tested extension template loading.

Reviewed at https://reviews.reviewboard.org/r/13499/