Add a simple CommaSeparatedValuesField for DB models.
We're running into some trouble that the old django-multiselectfield
project is unmaintained and does not support Django 4.x+. There's a
third-party fork which has had some updates, but no releases of that
that we can consume.

django-multiselectfield itself isn't very big, and of what it offers, we
don't actually use very much. This change adds a new field type to
Djblets to do the database serialization/deserialization to join and
split comma-separated values. It doesn't support the validation that
django-multiselectfield does, but we don't actually care about that
functionality.

Testing Done:
- Ran unit tests.
- Used this in Review Board and tested the functionality of the
  WebHookTarget admin form.

Reviewed at https://reviews.reviewboard.org/r/13560/