Fix a couple more places where we were using the old String.strip method.
Long ago, we monkey-patched the String class to include a `strip()`
method. This is now in the standard library as `trim()`, and we've
deprecated our version. This change fixes two places that still used it
that I had missed.

Testing Done:
Ran js-tests and saw that the deprecation warnings were no longer
showing up in the console.

Reviewed at https://reviews.reviewboard.org/r/13653/