Fix regressions in packaging extensions.
The rewrites for modern extension packaging had some regressions when
working off of a clean source tree and when processing lists of
entrypoints.

We had assumed the `build/lib` directory existed, but that wasn't a
guarantee at this point. This is now created when we need it.

We also assumed the extension's `static` directory existed, and broke if
it didn't. We now simply skip static media building in this case.

For the entrypoints, we now check if the result is a list or a string.
In either case, we convert to a string with an entrypoint per line, for
processing.

Testing Done:
Successfully built rbintegrations and Power Pack packages.

Reviewed at https://reviews.reviewboard.org/r/13762/