Begin moving Config Forms and related modules over to Ink styling.
One of the big inspirations for designing a new component library was
the mishmash of layout and styles that is the Config Forms suite of
components. This was an early attempt at standardizing the presentation
of forms, collections of forms, and fields, and we've learned a lot from
this. In time, the hope is to deprecate the vast majority of this module
in favor of Ink.

For now, we need to style this using Ink. This change begins that
process, moving much of the styling to Ink, and fixing up some small
layout issues in the process. This gets us dark mode and some
consistency in presentation.

We're also using `ink-c-button` now for action buttons in lists and
submit buttons in forms, and `ink-c-spinner` instead of
`djblets-o-spinner`.

Some related usage in the codebase (namely avatar forms) have been
updated as well.

This looks much as it did before on light mode, though we're making more
use of the accent color for selected navigation items. Dark mode looks
about as expected.

The Config Forms consumers often specialize much of the styling of
Config Forms, so changes will still be needed there.

Testing Done:
Unit tests pass.

Tested all of the My Account page on light and dark mode, with all
options available on all pages. There are still issues to address, but
they'll require further work in Review Board and in other modules in
Djblets.

Reviewed at https://reviews.reviewboard.org/r/13784/