Convert ExtensionManagerView to TypeScript/spina and fix bundles.
This change converts the ExtensionManagerView class to TypeScript and
spina.

This also fixes up some issues with our bundles in djblets. The
ExtensionManagerModel was converted to TypeScript, but inadvertently
included in the djblets-extensions bundle instead of the
djblets-extensions-admin bundle. In addition, we had neglected to remove
the other old filenames for the rest of the djblets-extensions bundle
contents from the list.

Testing Done:
- Ran js-tests.
- Used the extensions list in the Review Board admin and saw everything
  work correctly.

Reviewed at https://reviews.reviewboard.org/r/13803/