Remove deprecated-for-5.0 code from Djblets.
This change removes the rest of the APIs and usage that has been marked
as deprecated and removed for Djblets 5.0. This also turns on subquery
and join type checking by default for `assertQueries`, as was advertised
in the deprecation warnings for those. This required updating one unit
test for the join types, but otherwise works.

Testing Done:
Ran Djblets and Review Board unit tests.

Reviewed at https://reviews.reviewboard.org/r/13812/