Update datagrids for dark mode and modern styling.
Historically, Djblets's datagrid styles have been quite ugly and
somewhat broken. We've generally handled styling within Review Board
itself, placing styles there partly for convenience and partly to match
Review Board's look and feel. However, with Ink, we're now in a better
position to improve the default styling.

This change moves the styling to Ink, backports a lot of the Review
Board styles to Djblets, and introduces dark mode.

The styling is based on the (in-progress) Ink data table component. In
the future, the plan is to rebuild datagrids on top of a more stable
Ink foundation, but that's out of scope for now.

Testing Done:
Tested with Review Board, with an in-progress change that removes much
of the overridden styling in favor of Djblets.

This was tested on the dashboard, All Review Requests, Users, Groups,
and the user page.

Reviewed at https://reviews.reviewboard.org/r/13809/