Remove column-specific styles for datagrids.
This updates Djblets to remove the CSS for the age and summary columns.
These were provided in Review Board, but with some initial styling in
Djblets that Review Board had to override. Review Board will get some
styling updates to fill in the gaps removed from this CSS.

Testing Done:
Tested this along with the change for Review Board, verifying that styles
looked as expected.

Reviewed at https://reviews.reviewboard.org/r/13825/