Remove unnecessary and broken prototypeAttrs override for TableItemView.
TableItemView was including a prototypeAttrs override listing the
`template` member, but this is already included in the parent class one,
and we automerge those. Unfortunately, having this override seems to
have triggered a different bug in spina's automerge where we were losing
the other items. This was causing us to not have action handlers for the
extension list, breaking the "Enable" and "Disable" actions.

Testing Done:
Saw that I was able to enable and disable extensions.

Reviewed at https://reviews.reviewboard.org/r/13960/