Update integrations form and pop-up to work in dark mode.
The integrations form and pop-up are nominally part of the admin UI,
which doesn't yet support dark mode, but that's not always the case.
This change makes it so we use an Ink button and icon for the "Add
Integration" button, and styles the pop-up to be legible in dark mode.

Testing Done:
Used the integrations pop-up on a non-admin page with dark mode enabled.

Reviewed at https://reviews.reviewboard.org/r/13976/