Fix background color of last selected page entry in configforms.
When the last page on a config form page is selected, the background of
the selected item had a border radius that was too large. This change
fixes it up to use a border radius to fit correctly (the border radius
of the outer container minus the width of the border itself).

Testing Done:
Selected the last page in a configform set and saw that the TOC entry
looked correct.

Reviewed at https://reviews.reviewboard.org/r/13979/