Update configforms LESS variables to use ink constants.
This change updates a bunch of configforms variables to rely on Ink
instead of hard-coded values or legacy variables. This makes a lot of
configforms stuff look nicer, especially with dark mode. This also
switches the state icons for configforms list items from old fontawesome
versions to new ink ones.

We really need to give all of this a good scrubbing, but for now this is
a big improvement across the board.

Testing Done:
Flipped through all the pages of configforms we have (my account,
integrations UI in the RB admin, RBC team admin) and looked carefully at
everything.

Reviewed at https://reviews.reviewboard.org/r/13992/