Add typing for BaseRegistryHook and BaseRegistryMultiItemHook.
Registries gained support for typing a while back, helping to ensure
that any items provided or fetched were of compatible types.

This change extends that to `BaseRegistryHook` and
`BaseRegistryMultiItemHook`. This makes it easier to instantiate these
hooks with items and to assign a registry when defining a subclass.

Testing Done:
Unit tests pass.

Tested this with an in-progress hook definition.

Reviewed at https://reviews.reviewboard.org/r/13981/